Med_Cabinet/requirements.txt
pandas
numpy
requests
flask
flask_restful
json
pickle-mixin
sklearn.feature_extraction.text
re
string
os
sys
dotenv
psycopg2

spacy_install_virtualenv